fix integer to long

+++ modules/storm/src/main/java/in/dream_lab/bm/stream_iot/storm/sinks/Sink.java
@@ -47,7 +47,7 @@ public class Sink extends BaseRichBolt {
         //collector.emit(input,new Values(msgId));
         try {
         	//ba.batchLogwriter(System.currentTimeMillis(),msgId);
-            if (Integer.parseInt(msgId) % 5 == 0)
+            if (Long.parseLong(msgId) % 5 == 0)
 		        jr.batchWriter(System.currentTimeMillis(), msgId);
             // ba.batchLogwriter(System.currentTimeMillis(),msgId+","+exe_time);//addon
         } catch (Exception e) {
